refactor(#3116): Move static property utils
